chore(test-env): rename test env check fn
