+++ libs/ledger-live-common/src/exchange/swap/api/v5/fetchCurrencyFrom.ts
@@ -1,5 +1,5 @@
 import network from "@ledgerhq/live-network/network";
-import { isPlaywrightEnv } from "../../utils/isPlaywrightEnv";
+import { isIntegrationTestEnv } from "../../utils/isIntegrationTestEnv";
 import { getEnv } from "../../../../env";
 import { DEFAULT_SWAP_TIMEOUT_MS } from "../../const/timeout";
 
@@ -21,7 +21,7 @@ export async function fetchCurrencyFrom({
   currencyTo,
   additionalCoinsFlag = false,
 }: Props) {
-  if (isPlaywrightEnv()) return Promise.resolve(fetchCurrencyFrom);
+  if (isIntegrationTestEnv()) return Promise.resolve(fetchCurrencyFrom);
 
   const url = new URL(`${getEnv("SWAP_API_BASE_V5")}/currencies/from`);
   url.searchParams.append("providers-whitelist", providers.join(","));
