+++ libs/ledger-live-common/src/exchange/swap/api/v5/fetchCurrencyTo.ts
@@ -1,7 +1,7 @@
 import network from "@ledgerhq/live-network/network";
 import { getEnv } from "../../../../env";
 import { fetchCurrencyToMock } from "./__mocks__/fetchCurrencyTo.mocks";
-import { isPlaywrightEnv } from "../../utils/isPlaywrightEnv";
+import { isIntegrationTestEnv } from "../../utils/isIntegrationTestEnv";
 import { DEFAULT_SWAP_TIMEOUT_MS } from "../../const/timeout";
 import axios from "axios";
 import { LedgerAPI4xx } from "@ledgerhq/errors";
@@ -25,7 +25,7 @@ export async function fetchCurrencyTo({
   providers,
   additionalCoinsFlag = false,
 }: Props) {
-  if (isPlaywrightEnv()) return Promise.resolve(fetchCurrencyToMock);
+  if (isIntegrationTestEnv()) return Promise.resolve(fetchCurrencyToMock);
 
   const url = new URL(`${getEnv("SWAP_API_BASE_V5")}/currencies/to`);
