+++ libs/ledger-live-common/src/exchange/swap/getExchangeRates.ts
@@ -11,7 +11,7 @@ import {
 import { getProviderConfig, getSwapAPIBaseURL, getSwapAPIError } from "./";
 import { mockGetExchangeRates } from "./mock";
 import type { CustomMinOrMaxError, GetExchangeRates } from "./types";
-import { isPlaywrightEnv } from "./utils/isPlaywrightEnv";
+import { isIntegrationTestEnv } from "./utils/isIntegrationTestEnv";
 
 const getExchangeRates: GetExchangeRates = async ({
   exchange,
@@ -21,7 +21,7 @@ const getExchangeRates: GetExchangeRates = async ({
   timeout,
   timeoutErrorMessage,
 }) => {
-  if (isPlaywrightEnv()) return mockGetExchangeRates(exchange, transaction, currencyTo);
+  if (isIntegrationTestEnv()) return mockGetExchangeRates(exchange, transaction, currencyTo);
 
   const from = getAccountCurrency(exchange.fromAccount).id;
   const unitFrom = getAccountUnit(exchange.fromAccount);
