+++ libs/ledger-live-common/src/exchange/swap/getStatus.ts
@@ -2,10 +2,10 @@ import network from "@ledgerhq/live-network/network";
 import { getSwapAPIBaseURL } from "./";
 import { mockGetStatus } from "./mock";
 import type { GetMultipleStatus } from "./types";
-import { isPlaywrightEnv } from "./utils/isPlaywrightEnv";
+import { isIntegrationTestEnv } from "./utils/isIntegrationTestEnv";
 
 export const getMultipleStatus: GetMultipleStatus = async statusList => {
-  if (isPlaywrightEnv()) return mockGetStatus(statusList);
+  if (isIntegrationTestEnv()) return mockGetStatus(statusList);
 
   const res = await network({
     method: "POST",
