+++ libs/ledger-live-common/src/exchange/swap/postSwapState.ts
@@ -2,14 +2,14 @@ import network from "@ledgerhq/live-network/network";
 import { getSwapAPIBaseURL } from "./";
 import { mockPostSwapAccepted, mockPostSwapCancelled } from "./mock";
 import type { PostSwapAccepted, PostSwapCancelled } from "./types";
-import { isPlaywrightEnv } from "./utils/isPlaywrightEnv";
+import { isIntegrationTestEnv } from "./utils/isIntegrationTestEnv";
 
 export const postSwapAccepted: PostSwapAccepted = async ({
   provider,
   swapId = "",
   transactionId,
 }) => {
-  if (isPlaywrightEnv()) return mockPostSwapAccepted({ provider, swapId, transactionId });
+  if (isIntegrationTestEnv()) return mockPostSwapAccepted({ provider, swapId, transactionId });
 
   /**
    * Since swapId is requiered by the endpoit, don't call it if we don't have
@@ -33,7 +33,7 @@ export const postSwapAccepted: PostSwapAccepted = async ({
 };
 
 export const postSwapCancelled: PostSwapCancelled = async ({ provider, swapId = "" }) => {
-  if (isPlaywrightEnv()) return mockPostSwapCancelled({ provider, swapId });
+  if (isIntegrationTestEnv()) return mockPostSwapCancelled({ provider, swapId });
 
   /**
    * Since swapId is requiered by the endpoit, don't call it if we don't have
